web/requirements.txt
asgiref==3.2.3
attrs==19.3.0
Django==3.0.3
djangorestframework==3.11.0
importlib-metadata==1.5.0
more-itertools==8.2.0
packaging==20.1
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.6
pytest==5.3.5
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
wcwidth==0.1.8
zipp==3.0.0
psycopg2-binary==2.8.4
django-rest-swagger==2.2.0
djoser
pytest-django
django-filter

